streamlit_store_app/requirements.txt
streamlit==1.45.1
databricks-sdk==0.53.0
databricks-sql-connector==3.6.0
python-dotenv==1.1.0
pyyaml==6.0.1
mlflow==2.22.0
pandas>=2.2.0
plotly>=5.18.0
# setuptools>=68.0.0